// packages/auth/src/types/auth.d.ts
export declare const UserSchema: z.ZodObject<{
    id: z.ZodString;
    email: z.ZodString;
    name: z.ZodString;
    avatarUrl: z.ZodOptional<z.ZodString>;
    emailVerified: z.ZodOptional<z.ZodBoolean>;
    isVerified: z.ZodOptional<z.ZodBoolean>;
    createdAt: z.ZodString;
    updatedAt: z.ZodString;
    role: z.ZodString;
}, "strip", z.ZodTypeAny, {
    name?: string;
    email?: string;
    id?: string;
    role?: string;
    avatarUrl?: string;
    emailVerified?: boolean;
    isVerified?: boolean;
    createdAt?: string;
    updatedAt?: string;
}, {
    name?: string;
    email?: string;
    id?: string;
    role?: string;
    avatarUrl?: string;
    emailVerified?: boolean;
    isVerified?: boolean;
    createdAt?: string;
    updatedAt?: string;
}>;
export declare const AuthTokensSchema: z.ZodObject<{
    accessToken: z.ZodString;
    refreshToken: z.ZodString;
    expiresIn: z.ZodNumber;
}, "strip", z.ZodTypeAny, {
    accessToken?: string;
    refreshToken?: string;
    expiresIn?: number;
}, {
    accessToken?: string;
    refreshToken?: string;
    expiresIn?: number;
}>;
// ...
export interface User extends z.infer<typeof UserSchema> {
}
export interface AuthTokens extends z.infer<typeof AuthTokensSchema> {
}